requirement.txt
pip==25.0.1
aiohttp
python-dotenv
pytest==8.3.4
pytest-asyncio==0.23.1
requests